No hyphen in code-point* names
This has been a difficult decision. Properly 'code point' is written as
two words, as it is in the JDK APIs (.codePoints). However, with time
the spelling 'code-point*' feels more and more clunky or perhaps
pedantic to me. Clojure is generally freer or more organic in the choice
of names: a hyphen can be left out if a thing can be thought of as one
thing (like resultset-seq or macroexpand in core).

I'm not sure though, going with gut feeling. Drop the hyphen for now.